% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRcleanRatSquared.R
\name{ccr.run_complete}
\alias{ccr.run_complete}
\title{Title}
\usage{
ccr.run_complete(
  filename_single,
  min_reads_single = 30,
  EXPname,
  libraryAnnotation_single,
  display = FALSE,
  outdir = "./"
)
}
\arguments{
\item{filename_single}{}

\item{min_reads_single}{}

\item{EXPname}{}

\item{libraryAnnotation_single}{}

\item{display}{}

\item{outdir}{}
}
\value{

}
\description{
Title
}
