% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRcleanRatSquared.R
\name{ccr2.avgSingleGuides}
\alias{ccr2.avgSingleGuides}
\title{Title}
\usage{
ccr2.avgSingleGuides(dual_FC, single_FC, match_dual_single_seq, guide_id)
}
\arguments{
\item{dual_FC}{}

\item{single_FC}{}

\item{match_dual_single_seq}{}

\item{guide_id}{}
}
\description{
collapse avgFC via mean for each gene in guide1 or guide2 across all the other genes
store also number of genes from which mean is computed
}
