% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRcleanRatSquared.R
\name{ccr2.compute_bliss}
\alias{ccr2.compute_bliss}
\title{Title}
\usage{
ccr2.compute_bliss(dual_FC, corrected = FALSE)
}
\arguments{
\item{dual_FC}{}

\item{corrected}{}
}
\value{

}
\description{
Title
}
