% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRcleanRatSquared.R
\name{ccr2.createPseudoSingle}
\alias{ccr2.createPseudoSingle}
\title{Title}
\usage{
ccr2.createPseudoSingle(dual_FC, guide_id)
}
\arguments{
\item{dual_FC}{}

\item{guide_id}{}
}
\description{
psuedo single avgFC via mean for each gene in guide1 or guide2 across all the other genes
store also number of genes from which mean is computed
}
