% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRcleanRatSquared.R
\name{ccr2.logFCs2chromPos}
\alias{ccr2.logFCs2chromPos}
\title{Title}
\usage{
ccr2.logFCs2chromPos(dual_FC, dual_library)
}
\arguments{
\item{dual_FC}{}

\item{dual_library}{}
}
\description{
Title
}
