% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRcleanRatSquared.R
\name{ccr2.match_singletons_singleFC}
\alias{ccr2.match_singletons_singleFC}
\title{Title}
\usage{
ccr2.match_singletons_singleFC(
  match_dual_single_seq,
  single_FC,
  singletons_summary_FC,
  saveToFig = FALSE,
  display = TRUE,
  saveFormat = NULL,
  EXPname = "",
  outdir = "./"
)
}
\arguments{
\item{match_dual_single_seq}{}

\item{single_FC}{}

\item{singletons_summary_FC}{}

\item{saveToFig}{}

\item{display}{}

\item{saveFormat}{}

\item{EXPname}{}

\item{outdir}{}
}
\description{
Title
}
