% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRcleanRatSquared.R
\name{ccr2.modelSingleVSCombined}
\alias{ccr2.modelSingleVSCombined}
\title{Title}
\usage{
ccr2.modelSingleVSCombined(
  dual_collapsed,
  single_FC,
  guide_id,
  correctGW = NULL,
  display = TRUE,
  saveToFig = FALSE,
  saveFormat = "pdf",
  outdir = "./",
  EXPname = ""
)
}
\arguments{
\item{dual_collapsed}{}

\item{single_FC}{}

\item{guide_id}{}

\item{correctGW}{}

\item{display}{}

\item{saveToFig}{}

\item{saveFormat}{}

\item{outdir}{}

\item{EXPname}{}
}
\description{
Model to convert combined dual to single screens
that depends on the specific gene selection
}
