% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRcleanRatSquared.R
\name{ccr2.plot_bliss_vs_FC}
\alias{ccr2.plot_bliss_vs_FC}
\title{Title}
\usage{
ccr2.plot_bliss_vs_FC(
  dual_FC,
  corrected = FALSE,
  THR_FC = -1,
  THR_BLISS = -1,
  saveToFig = FALSE,
  saveFormat = "pdf",
  outdir = "./",
  EXPname = ""
)
}
\arguments{
\item{dual_FC}{}

\item{corrected}{}

\item{THR_FC}{}

\item{THR_BLISS}{}

\item{saveToFig}{}

\item{saveFormat}{}

\item{outdir}{}

\item{EXPname}{}
}
\value{

}
\description{
Title
}
