% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRcleanRatSquared.R
\name{ccr2.plotClasses}
\alias{ccr2.plotClasses}
\title{Plot logFC distributions for before and after CRISPRcleanR^2 correction, separating per dual guides classes}
\usage{
ccr2.plotClasses(
  dual_FC_correctedFC,
  saveToFig = F,
  saveFormat = "pdf",
  outdir = "./",
  EXPname = ""
)
}
\arguments{
\item{dual_FC_correctedFC}{Data frame containing information about corrected logFC values for dual KO.}

\item{saveToFig}{Logical, indicating whether to save the generated plot to a file. Default is FALSE.}

\item{saveFormat}{File format for saving the plot (e.g., ".pdf", ".png"). Default is ".pdf".}

\item{outdir}{Directory to save the plot. Default is "./".}

\item{EXPname}{Experiment name to include in the plot filename.}
}
\description{
This function generates boxplots and density plots for the distribution of logFC values (before and after CRISPRcleanR^2 correction),
comparing dual guides classes.
}
