% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRcleanRatSquared.R
\name{ccr2.plotCNA}
\alias{ccr2.plotCNA}
\title{Title}
\usage{
ccr2.plotCNA(
  dual_FC_correctedFC,
  dual_CNA,
  saveToFig = FALSE,
  saveFormat = "pdf",
  outdir = "./",
  EXPname = "",
  excludeGene = NULL
)
}
\arguments{
\item{dual_FC_correctedFC}{}

\item{dual_CNA}{}

\item{saveToFig}{}

\item{saveFormat}{}

\item{outdir}{}

\item{EXPname}{}

\item{excludeGene}{}
}
\description{
plot: distribution with respect to CN
}
