% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRcleanRatSquared.R
\name{ccr2.plotMatchingSingle}
\alias{ccr2.plotMatchingSingle}
\title{Title}
\usage{
ccr2.plotMatchingSingle(
  dual_FC_correctedFC,
  match_dual_single_seq,
  single_correctedFCs,
  dual_CNA,
  saveToFig = FALSE,
  saveFormat = "pdf",
  outdir = "./",
  EXPname = ""
)
}
\arguments{
\item{dual_FC_correctedFC}{}

\item{match_dual_single_seq}{}

\item{single_correctedFCs}{}

\item{dual_CNA}{}

\item{saveToFig}{}

\item{saveFormat}{}

\item{outdir}{}

\item{EXPname}{}
}
\description{
plot: before and after correction of matching single dual gene
}
