% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRcleanRatSquared.R
\name{ccr2.run}
\alias{ccr2.run}
\title{Run CRISPRcleanR^2 analysis for dual KO screen ("double-cut" pairs).}
\usage{
ccr2.run(
  dual_FC,
  single_correctedFCs,
  libraryAnnotation_dual,
  match_dual_single_seq,
  EXPname = "",
  saveToFig = FALSE,
  display = TRUE,
  saveFormat = NULL,
  outdir = "./",
  correctGW,
  ...
)
}
\arguments{
\item{dual_FC}{The data frame containing dual KO logFCs (output of ccr2.logFCs2chromPos).}

\item{single_correctedFCs}{The data frame containing corrected single KO logFCs already corrected via CRISPRcleanR (output of ccr.run_complete).}

\item{libraryAnnotation_dual}{Data frame containing information from the dual library (one row per guide pair).}

\item{match_dual_single_seq}{A data frame with one row per sgRNA in dual library (union of position 1 and position 2), including matched single library IDs and sequences (output of ccr2.matchDualandSingleSeq).}

\item{EXPname}{A character string specifying the experiment name (default is "").}

\item{saveToFig}{Logical, whether to save figures (default is FALSE).}

\item{display}{Logical, whether to display figures (default is TRUE).}

\item{saveFormat}{The format in which to save figures (default is NULL).}

\item{outdir}{The directory to save figures (default is "./").}

\item{correctGW}{A character specifying the type of pseudo-single logFCs mean centering after data-injection ("CHR": per chromosome, "GW": genome-wide, or NULL for no centering). See ccr2.injectData.}

\item{...}{Additional parameters to be passed.}
}
\value{
A list containing maximum and minimum corrections allowed, model performance and estimated
for single VS pseudo-single, identified pseudo-single segments and correction, and final dual logFC data frame with CRISPRcleanR^2 correction.
}
\description{
This function performs a comprehensive CRISPRcleanR^2 analysis in dual KO screen for "double-cut" pairs,
meaning both guides in position1 and position 2 target an existing genomic region. The following steps are performed:
\itemize{
\item creation of pseudo single scores from dual KO logFCs,
\item modeling of single KO logFCs vs. pseudo single logFCs for matching guides,
\item injection of data into pseudo single logFCs space from genome-wide single KO screen,
\item application of CRISPRcleanR to pseudo-single logFCs (genome-wide from original plus injected data),
\item approximate linear systems with box constraints to obtain pairwise correction from pseudo-single correction.
It returns information such as the maximum and minimum corrections allowed,
model performance for single VS pseudo-single,
identified pseudo-single segments and correction, and
final dual logFC data frame with CRISPRcleanR^2 correction.
}
}
\seealso{
\code{\link{ccr2.createPseudoSingle_combine}},
\code{\link{ccr2.modelSingleVSPseudoSingle}},
\code{\link{ccr2.injectData}},
\code{\link{ccr.GWclean}},
\code{\link{ccr2.filterGWclean}},
\code{\link{ccr2.solveLinearSys}},
\code{\link{ccr2.logFCs2chromPos}},
\code{\link{ccr.run_complete}},
\code{\link{ccr2.matchDualandSingleSeq}}
}
