% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRcleanRatSquared.R
\name{ccr2.run_nontarget}
\alias{ccr2.run_nontarget}
\title{Title}
\usage{
ccr2.run_nontarget(
  dual_FC,
  single_correctedFCs,
  libraryAnnotation_dual,
  match_dual_single_seq,
  EXPname = "",
  saveToFig = FALSE,
  display = TRUE,
  saveFormat = NULL,
  outdir = "./",
  weighted_reg = TRUE,
  correctGW
)
}
\arguments{
\item{dual_FC}{}

\item{single_correctedFCs}{}

\item{libraryAnnotation_dual}{}

\item{match_dual_single_seq}{}

\item{EXPname}{}

\item{saveToFig}{}

\item{display}{}

\item{saveFormat}{}

\item{outdir}{}

\item{weighted_reg}{}

\item{correctGW}{}
}
\value{

}
\description{
Title
}
