% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRcleanRatSquared.R
\name{ccr2.solveLinearSys}
\alias{ccr2.solveLinearSys}
\title{Title}
\usage{
ccr2.solveLinearSys(
  dual_FC,
  dual_collapsedG1_correctedFCs,
  dual_collapsedG2_correctedFCs,
  display = TRUE,
  saveToFig = FALSE,
  saveFormat = "pdf",
  outdir = "./",
  EXPname = ""
)
}
\arguments{
\item{dual_FC}{}

\item{dual_collapsedG1_correctedFCs}{}

\item{dual_collapsedG2_correctedFCs}{}

\item{display}{}

\item{saveToFig}{}

\item{saveFormat}{}

\item{outdir}{}

\item{EXPname}{}
}
\description{
Title
}
