% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRcleanRatSquared.R
\name{ccr2.add_CNA}
\alias{ccr2.add_CNA}
\title{Title}
\usage{
ccr2.add_CNA(CNA, dual_FC)
}
\arguments{
\item{CNA}{}

\item{dual_FC}{}
}
\value{

}
\description{
combine copy number info with dual FC data.frame
}
