% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRcleanRatSquared.R
\name{ccr2.createPseudoSingle_combine}
\alias{ccr2.createPseudoSingle_combine}
\title{Title}
\usage{
ccr2.createPseudoSingle_combine(
  dual_FC,
  saveToFig = FALSE,
  display = TRUE,
  saveFormat = NULL,
  EXPname = "",
  outdir = "./"
)
}
\arguments{
\item{dual_FC}{}

\item{saveToFig}{}

\item{display}{}

\item{saveFormat}{}

\item{EXPname}{}

\item{outdir}{}
}
\description{
Title
}
