% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRcleanRatSquared.R
\name{ccr2.filterGWclean}
\alias{ccr2.filterGWclean}
\title{Title}
\usage{
ccr2.filterGWclean(
  dataInjection_correctedFCs,
  dual_collapsed,
  guide_id,
  saveToFig = FALSE,
  display = TRUE,
  saveFormat = ".pdf",
  outdir = "./",
  EXPname = ""
)
}
\arguments{
\item{dataInjection_correctedFCs}{}

\item{dual_collapsed}{}

\item{guide_id}{}

\item{saveToFig}{}

\item{display}{}

\item{saveFormat}{}

\item{outdir}{}

\item{EXPname}{}
}
\description{
adjust ccr output applied to collapsed and filter per considered guide in dual
}
