% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRcleanRatSquared.R
\name{ccr2.get_summary_singletons}
\alias{ccr2.get_summary_singletons}
\title{Title}
\usage{
ccr2.get_summary_singletons(dual_FC, corrected = F)
}
\arguments{
\item{dual_FC}{}

\item{corrected}{}
}
\description{
Title
}
