% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRcleanRatSquared.R
\name{ccr2.injectData}
\alias{ccr2.injectData}
\title{Title}
\usage{
ccr2.injectData(
  model_single_to_pseudo,
  pseudo_single_FC,
  single_FC,
  guide_id,
  correctGW = NULL
)
}
\arguments{
\item{model_single_to_pseudo}{}

\item{pseudo_single_FC}{}

\item{single_FC}{}

\item{guide_id}{}

\item{correctGW}{}
}
\description{
Title
}
