% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRcleanRatSquared.R
\name{ccr2.matchDualandSingleSeq}
\alias{ccr2.matchDualandSingleSeq}
\title{Title}
\usage{
ccr2.matchDualandSingleSeq(dual_library, single_library)
}
\arguments{
\item{dual_library}{}

\item{single_library}{}
}
\description{
Title
}
