% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRcleanRatSquared.R
\name{ccr2.NormfoldChanges}
\alias{ccr2.NormfoldChanges}
\title{Title}
\usage{
ccr2.NormfoldChanges(Dframe, min_reads = 30)
}
\arguments{
\item{Dframe}{}

\item{min_reads}{}
}
\value{

}
\description{
Title
}
