% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRcleanRatSquared.R
\name{ccr2.plot_correction}
\alias{ccr2.plot_correction}
\title{Title}
\usage{
ccr2.plot_correction(
  df,
  saveToFig = FALSE,
  saveFormat = "pdf",
  outdir = "./",
  EXPname = ""
)
}
\arguments{
\item{df}{}

\item{saveToFig}{}

\item{saveFormat}{}

\item{outdir}{}

\item{EXPname}{}
}
\value{

}
\description{
Title
}
