% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRcleanRatSquared.R
\name{ccr2.plotClasses}
\alias{ccr2.plotClasses}
\title{Title}
\usage{
ccr2.plotClasses(
  dual_FC_correctedFC,
  saveToFig = F,
  saveFormat = "pdf",
  outdir = "./",
  EXPname = ""
)
}
\arguments{
\item{dual_FC_correctedFC}{}

\item{saveToFig}{}

\item{saveFormat}{}

\item{outdir}{}

\item{EXPname}{}
}
\description{
plot: distribution before and after based on classes
}
