% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRcleanRatSquared.R
\name{ccr2.plotCNAdensity}
\alias{ccr2.plotCNAdensity}
\title{Title}
\usage{
ccr2.plotCNAdensity(
  dual_FC_correctedFC,
  dual_CNA,
  saveToFig = F,
  saveFormat = "pdf",
  outdir = "./",
  EXPname = "",
  excludeGene = NULL,
  CN_thr,
  var_to_plot = "observed"
)
}
\arguments{
\item{dual_FC_correctedFC}{}

\item{dual_CNA}{}

\item{saveToFig}{}

\item{saveFormat}{}

\item{outdir}{}

\item{EXPname}{}

\item{excludeGene}{}

\item{CN_thr}{}

\item{var_to_plot}{}
}
\description{
plot: density divided per max CNA threshold
}
