% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRcleanRatSquared.R
\name{ccr2.run_complete}
\alias{ccr2.run_complete}
\title{Title}
\usage{
ccr2.run_complete(
  filename_single,
  min_reads_single = 30,
  libraryAnnotation_single,
  min_reads = 30,
  libraryAnnotation_dual,
  dual_count,
  EXPname,
  display = FALSE,
  outdir = "./",
  saveToFig,
  saveFormat = "pdf",
  correctGW,
  excludeGene_plot = NULL,
  CNA,
  CN_thr = 8,
  ...
)
}
\arguments{
\item{filename_single}{}

\item{min_reads_single}{}

\item{libraryAnnotation_single}{}

\item{min_reads}{}

\item{libraryAnnotation_dual}{}

\item{dual_count}{}

\item{EXPname}{}

\item{display}{}

\item{outdir}{}

\item{saveToFig}{}

\item{saveFormat}{}

\item{correctGW}{}

\item{excludeGene_plot}{}

\item{CNA}{}

\item{CN_thr}{}

\item{...}{}
}
\value{

}
\description{
Title
}
