% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRcleanRatSquared.R
\name{ccr2.solveLinearSys}
\alias{ccr2.solveLinearSys}
\title{Title}
\usage{
ccr2.solveLinearSys(
  dual_FC,
  pseudo_single_p1_correctedFCs,
  pseudo_single_p2_correctedFCs,
  split_graph = TRUE,
  display = TRUE,
  saveToFig = FALSE,
  saveFormat = "pdf",
  outdir = "./",
  EXPname = ""
)
}
\arguments{
\item{dual_FC}{}

\item{pseudo_single_p1_correctedFCs}{}

\item{pseudo_single_p2_correctedFCs}{}

\item{split_graph}{}

\item{display}{}

\item{saveToFig}{}

\item{saveFormat}{}

\item{outdir}{}

\item{EXPname}{}
}
\description{
Title
}
