% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRcleanRatSquared.R
\name{get_input_data}
\alias{get_input_data}
\title{Load matched files for dual KO}
\usage{
get_input_data(param_file)
}
\arguments{
\item{param_file}{.tsv file, the first column indicates the type of data, the
second the file location. Can include multiple files that contain the name
count or library BUT the lenght must coincide.}
}
\value{
a list * CNA: copy number, one row per gene * count:  dual KO count,
one row per guides combination * library: metadata for dual KO library *
CL_name: cell line name * out_fold: location to store results
}
\description{
get_input_data() load files as indicated in param_file. Must include count,
library and copy_number fields.
}
