% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRcleanRatSquared.R
\name{get_input_data.v1}
\alias{get_input_data.v1}
\title{Load matched files for dual KO}
\usage{
get_input_data.v1(param_file = NULL, param_list = NULL)
}
\arguments{
\item{param_file}{}

\item{param_list}{}
}
\value{

}
\description{
get_input_data.v1() load files as indicated in param_file or param_list.
Must include result_file, library_file and copy_number_file.
Only one file per type is passed (suitable for different batches combined)
Additional used inputs are input_fold and out_fold (default = "./") and CL_name (default = "./")
}
