% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRcleanRatSquared.R
\name{get_pairwise_correction}
\alias{get_pairwise_correction}
\title{Title}
\usage{
get_pairwise_correction(
  matrix_interactions,
  combinations,
  dual_FC,
  pseudo_single_p1,
  pseudo_single_p2,
  max_corr
)
}
\arguments{
\item{matrix_interactions}{}

\item{combinations}{}

\item{dual_FC}{}

\item{pseudo_single_p1}{}

\item{pseudo_single_p2}{}

\item{max_corr}{}
}
\value{

}
\description{
Title
}
