% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRcleanRatSquared.R
\name{match_dual_to_single}
\alias{match_dual_to_single}
\title{Title}
\usage{
match_dual_to_single(dual_seq, single_library_seq)
}
\arguments{
\item{dual_seq}{}

\item{single_library_seq}{}
}
\description{
Title
}
